backend/pyproject.toml
[tool.poetry]
name = "magacin-backend"
version = "0.1.0"
description = "Backend monorepo for Interni WMS Operativni Sloj"
authors = ["Cungu & Co Tech Team"]
packages = [
    { include = "app_common" },
    { include = "services", from = "services" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
httpx = "^0.27.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
redis = "^5.0.4"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pendulum = "^3.0.0"
structlog = "^24.1.0"
uvloop = "^0.19.0"
watchfiles = "^0.21.0"
pdfplumber = "^0.11.4"
openpyxl = "^3.1.2"
python-multipart = "^0.0.9"
python-socketio = "^5.11.1"
prometheus-fastapi-instrumentator = "^7.0.0"
prometheus-client = "^0.20.0"
croniter = "^1.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = { extras = ["cli"], version = "^0.27.0" }
black = "^24.4.0"
isort = "^5.13.2"
ruff = "^0.4.5"
mypy = "^1.10.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
